# openapi.yaml
openapi: 3.1.0
info:
  title: Fusion E-Commerce Backend APIs
  description: |
    API documentation for the Fusion E-Commerce backend server. This documentation provides detailed information on all available endpoints for managing products, users, authentication, and more.
  version: 1.1.0
  contact:
    email: support@fusion-ecommerce.com
servers:
  - url: https://fusion-electronics-api.vercel.app/
    description: Production server
  - url: https://fusion-electronics-api.vercel.app
    description: Local development server
paths:
...
  /api/products/{id}:
    get:
      summary: Get a product by id
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
  /api/products/category/{category}:
    get:
      summary: Get products by category
      tags: [Products]
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Products in category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/products/{id}/rating:
    put:
      summary: Update the product rating
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: number
                  format: float
      responses:
        '200':
          description: Product rating updated successfully
        '400':
          description: Bad request
...
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        rating:
          type: number
...